+++ refinedstorage2-platform-api/src/main/java/com/refinedmods/refinedstorage2/platform/api/resource/FuzzyModeNormalizer.java
@@ -3,8 +3,8 @@ package com.refinedmods.refinedstorage2.platform.api.resource;
 public interface FuzzyModeNormalizer<T> {
     T normalize();
 
-    static Object tryNormalize(boolean exactMode, Object value) {
-        if (exactMode) {
+    static Object tryNormalize(boolean fuzzyMode, Object value) {
+        if (!fuzzyMode) {
             return value;
         }
         if (value instanceof FuzzyModeNormalizer<?> normalizer) {