+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/containermenu/storage/StorageSettingsProvider.java
@@ -13,9 +13,9 @@ public interface StorageSettingsProvider {
 
     void setFilterMode(FilterMode filterMode);
 
-    boolean isExactMode();
+    boolean isFuzzyMode();
 
-    void setExactMode(boolean exactMode);
+    void setFuzzyMode(boolean fuzzyMode);
 
     AccessMode getAccessMode();